% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iccs_2009.R
\docType{data}
\name{iccs_2009}
\alias{iccs_2009}
\title{ICCS 2009}
\format{
Object class \code{"cross"}; ver \code{\link[qtl]{read.cross}}.
}
\source{
IEA, <https://www.iea.nl/index.php/data-tools/repository/iccs>
}
\usage{
data(iccs_2009)
}
\description{
Merge of students data from the International Civic and Citizenship Education Study 2009 (ICCS 2009).
The study contains representative samples of 8th graders (14 years of age), from 38 different educational systems.
}
\examples{
dplyr::glimpse(ilsa::iccs_2009)
}
\references{
Schulz, W., Ainley, J., & Fraillon, J. (2011). ICCS 2009 Technical Report (W. Schulz, J. Ainley, & J. Fraillon (eds.)). International Association for the Evaluation of Educational Achievement (IEA). http://www.iea.nl/fileadmin/user_upload/Publications/Electronic_versions/ICCS_2009_Technical_Report.pdf
}
\keyword{datasets}
