% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iccs_2016.R
\docType{data}
\name{iccs_2016}
\alias{iccs_2016}
\title{ICCS 2016}
\format{
Object class \code{"cross"}; ver \code{\link[qtl]{read.cross}}.
}
\source{
IEA, <https://www.iea.nl/index.php/data-tools/repository/iccs>
}
\usage{
data(iccs_2016)
}
\description{
Merge of students data from the International Civic and Citizenship Education Study 2016 (ICCS 2016).
The study contains representative samples of 8th graders (14 years of age), from 24 different educational systems.
}
\examples{
dplyr::glimpse(ilsa::iccs_2016)
}
\references{
Schulz, W., Carstens, R., Losito, B., & Fraillon, J. (2018). ICCS 2016 Technical Report (W. Schulz, R. Carstens, B. Losito, & J. Fraillon (eds.)). International Association for the Evaluation of Educational Achievement (IEA).
}
\keyword{datasets}
